+++ src/main/java/com/project/real_estate_project03_team02/entity/concretes/Category.java
@@ -1,4 +1,4 @@
-package com.project.real_estate_project03_team02.entity;
+package com.project.real_estate_project03_team02.entity.concretes;
