+++ src/main/java/com/project/real_estate_project03_team02/entity/concretes/CategoryPropertyKey.java
@@ -1,4 +1,5 @@
-package com.project.real_estate_project03_team02.entity;
+package com.project.real_estate_project03_team02.entity.concretes;
+
 
 
 import lombok.*;
